There were erros with the EEPROM memory. I incorrectly numbered the addresses. The numbered addresses have al been replaced by for-loops and calculations.
There is still a bug remaining. During a program you can adjust the servo positions and those are retained. However I believe that the eeprom address is not correct.
I use the relays to show which address is set, but all LEDs burn.

The difference between F9 and F10 is always clear, no bug there.