Altered code so that servo's can be adjusted with F11 and F12 respectively. This prevents accidental presses on F9 and F10. One is forced to hold down the shift button to adjust servo positions.
I want to add a system with hardcoded routes ala NX style. There are 9 avaiable tracks with many differenct combinations. I also want to the relays to switch on for a route. All relays should switch off when a new route is picked. The controller cannot always know when the train has reached it's position. This is only possible if a track has diode catches so the controller knows by the lack of current draw that a train has stopped